+++ datumaro/util/test_utils.py
@@ -92,7 +92,7 @@ def compare_datasets(test, expected, actual, ignored_attrs=None,
             x.subset == item_a.subset)
         test.assertFalse(item_b is None, item_a.id)
         test.assertEqual(item_a.attributes, item_b.attributes)
-        if require_images or \
+        if (require_images and item_a.has_image and item_a.image.has_data) or \
                 item_a.has_image and item_a.image.has_data and \
                 item_b.has_image and item_b.image.has_data:
             test.assertEqual(item_a.image, item_b.image, item_a.id)